+++ docs/src/calling-contracts/other-contracts.md
@@ -3,5 +3,5 @@
 Your contract method might be calling other contracts. To do so, you must feed the external contract IDs that your contract depends on to the method you're calling. You do it by chaining `.set_contracts(&[external_contract_id, ...])` to the method you want to call. For instance:
 
 ```rust,ignore
-{{#include ../../../packages/fuels/tests/behaviors/mod.rs:external_contract}}
+{{#include ../../../packages/fuels/tests/contracts/mod.rs:external_contract}}
 ```
