+++ docs/src/contracts/the-setup-contract-test-macro.md
@@ -25,7 +25,7 @@ The input of the macro are the contract instance variable name, wallet variable 
 If you want to deploy contracts to the same provider, you have to set the wallet name of the first macro to `wallet` and all the remaining wallet names to `None`. The first macro will create `wallet` and bring it into context, and the other macros will use it instead of creating new ones. Let's see it in an example.
 
 ```rust,ignore
-{{#include ../../../packages/fuels/tests/contract_instance/mod.rs:contract_setup_macro_multi}}
+{{#include ../../../packages/fuels/tests/contracts//mod.rs:contract_setup_macro_multi}}
 ```
 
 In this example, three contracts are deployed on the same provider using the `wallet`. The second and third macro use the same contract but have different IDs because of the deployment with salt. Both of them can call the first contract by using its ID.
@@ -33,5 +33,5 @@ In this example, three contracts are deployed on the same provider using the `wa
 In addition, you can manually create the `wallet` variable and then use it inside the macro. This is useful if you want to create custom wallets or providers, but still want to use the macro to reduce boilerplate code. Below is an example of this approach.
 
 ```rust,ignore
-{{#include ../../../packages/fuels/tests/contract_instance/mod.rs:contract_setup_macro_manual_wallet}}
+{{#include ../../../packages/fuels/tests/contracts/mod.rs:contract_setup_macro_manual_wallet}}
 ```
