+++ examples/contracts/src/lib.rs
@@ -29,7 +29,7 @@ mod tests {
         abigen!(
             MyContract,
             // This path is relative to the workspace (repository) root
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         // This helper will launch a local node and provide a test wallet linked to it
@@ -44,7 +44,7 @@ mod tests {
         // be used to initialize the instance
         let contract_id = Contract::deploy(
             // This path is relative to the current crate (examples/contracts)
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::new(Some(gas_price), Some(gas_limit), Some(maturity)),
             StorageConfiguration::default(),
@@ -86,7 +86,7 @@ mod tests {
         setup_contract_test!(
             contract_instance,
             wallet,
-            "packages/fuels/tests/behaviors/contract_test"
+            "packages/fuels/tests/contracts/contract_test"
         );
 
         let response = contract_instance
@@ -107,13 +107,13 @@ mod tests {
 
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let wallet = launch_provider_and_get_wallet().await;
 
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -145,14 +145,14 @@ mod tests {
         // ANCHOR: abigen_example
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
         // ANCHOR_END: abigen_example
 
         let wallet = launch_provider_and_get_wallet().await;
 
         let contract_id_1 = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -165,7 +165,7 @@ mod tests {
         let salt: [u8; 32] = rng.gen();
 
         let contract_id_2 = Contract::deploy_with_parameters(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -187,13 +187,13 @@ mod tests {
 
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let wallets = launch_custom_provider_and_get_wallets(WalletsConfig::default(), None).await;
 
         let contract_id_1 = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallets[0],
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -213,7 +213,7 @@ mod tests {
         assert_eq!(42, response.value);
 
         let contract_id_2 = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallets[1],
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -241,12 +241,12 @@ mod tests {
         use fuels::prelude::*;
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let wallet = launch_provider_and_get_wallet().await;
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -319,12 +319,12 @@ mod tests {
         use fuels::prelude::*;
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/token_ops/out/debug/token_ops-abi.json"
+            "packages/fuels/tests/contracts/token_ops/out/debug/token_ops-abi.json"
         );
 
         let wallet = launch_provider_and_get_wallet().await;
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/token_ops/out/debug/token_ops\
+            "../../packages/fuels/tests/contracts/token_ops/out/debug/token_ops\
         .bin",
             &wallet,
             TxParameters::default(),
@@ -358,11 +358,11 @@ mod tests {
         use fuels::tx::Receipt;
         abigen!(
             TestContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
         let wallet = launch_provider_and_get_wallet().await;
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test\
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test\
         .bin",
             &wallet,
             TxParameters::default(),
@@ -402,7 +402,7 @@ mod tests {
         // Replace with your contract ABI.json path
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
         let wallet = launch_provider_and_get_wallet().await;
         // Your contract ID as a String.
@@ -421,13 +421,13 @@ mod tests {
         use fuels::prelude::*;
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let wallet = launch_provider_and_get_wallet().await;
 
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -460,13 +460,13 @@ mod tests {
 
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let wallet = launch_provider_and_get_wallet().await;
 
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -509,13 +509,13 @@ mod tests {
 
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let wallet = launch_provider_and_get_wallet().await;
 
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet,
             TxParameters::default(),
             StorageConfiguration::default(),
@@ -551,7 +551,7 @@ mod tests {
         use fuels::prelude::*;
         abigen!(
             MyContract,
-            "packages/fuels/tests/behaviors/contract_test/out/debug/contract_test-abi.json"
+            "packages/fuels/tests/contracts/contract_test/out/debug/contract_test-abi.json"
         );
 
         let config = WalletsConfig::new(Some(2), Some(1), Some(DEFAULT_COIN_AMOUNT));
@@ -560,7 +560,7 @@ mod tests {
         let wallet_2 = wallets.pop().unwrap();
 
         let contract_id = Contract::deploy(
-            "../../packages/fuels/tests/behaviors/contract_test/out/debug/contract_test.bin",
+            "../../packages/fuels/tests/contracts/contract_test/out/debug/contract_test.bin",
             &wallet_1,
             TxParameters::default(),
             StorageConfiguration::default(),
