--- packages/fuels/tests/contract_instance/collision_in_fn_names/Forc.toml
@@ -1,7 +1,0 @@
-[project]
-authors = ["Fuel Labs <contact@fuel.sh>"]
-entry = "main.sw"
-license = "Apache-2.0"
-name = "collision_in_fn_names"
-
-[dependencies]
