--- packages/fuels/tests/contract_instance/collision_in_fn_names/src/main.sw
@@ -1,11 +1,0 @@
-contract;
-
-abi MyContract {
-    fn new() -> u64;
-}
-
-impl MyContract for Contract {
-    fn new() -> u64 {
-        12345u64
-    }
-}
