+++ packages/fuels/tests/contracts/auth_testing_abi/Forc.toml
