+++ packages/fuels/tests/contracts/auth_testing_abi/src/main.sw
