+++ packages/fuels/tests/contracts/auth_testing_contract/Forc.toml
