+++ packages/fuels/tests/contracts/auth_testing_contract/src/main.sw
