+++ packages/fuels/tests/contracts/contract_test/Forc.toml
