+++ packages/fuels/tests/contracts/contract_test/src/main.sw
@@ -34,6 +34,7 @@ abi TestContract {
     fn array_of_enums(p: [State; 2]) -> [State; 2];
     fn get_array(p: [u64; 2]) -> [u64; 2];
     fn get_msg_amount() -> u64;
+    fn new() -> u64;
 }
 
 const COUNTER_KEY = 0x0000000000000000000000000000000000000000000000000000000000000000;
@@ -85,5 +86,9 @@ impl TestContract for Contract {
 
     fn get_array(p: [u64; 2]) -> [u64; 2] {
         p
+    }
+
+    fn new() -> u64 {
+        12345u64
     }
 }
