+++ packages/fuels/tests/contracts/foo_caller_contract/Forc.toml
