+++ packages/fuels/tests/contracts/foo_caller_contract/src/main.sw
