+++ packages/fuels/tests/contracts/foo_contract/Forc.toml
