+++ packages/fuels/tests/contracts/foo_contract/src/main.sw
