+++ packages/fuels/tests/contracts/foo_contract_abi/Forc.toml
