+++ packages/fuels/tests/contracts/foo_contract_abi/src/main.sw
