+++ packages/fuels/tests/contracts/large_return_data/src/main.sw
