+++ packages/fuels/tests/contracts/library_test/Forc.toml
