+++ packages/fuels/tests/contracts/library_test/src/main.sw
