+++ packages/fuels/tests/contracts/liquidity_pool/Forc.toml
