+++ packages/fuels/tests/contracts/liquidity_pool/src/main.sw
