+++ packages/fuels/tests/contracts/mod.rs
@@ -1,4 +1,3 @@
-use fuel_gql_client::fuel_tx::Receipt;
 use fuels::prelude::*;
 
 #[tokio::test]
@@ -6,7 +5,7 @@ async fn test_multiple_args() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/contract_test"
+        "packages/fuels/tests/contracts/contract_test"
     );
 
     // Make sure we can call the contract with multiple arguments
@@ -31,7 +30,7 @@ async fn test_contract_calling_contract() -> Result<(), Error> {
     setup_contract_test!(
         foo_contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/foo_contract"
+        "packages/fuels/tests/contracts/foo_contract"
     );
     let foo_contract_id = foo_contract_instance.get_contract_id();
 
@@ -43,7 +42,7 @@ async fn test_contract_calling_contract() -> Result<(), Error> {
     setup_contract_test!(
         foo_caller_contract_instance,
         None,
-        "packages/fuels/tests/behaviors/foo_caller_contract"
+        "packages/fuels/tests/contracts/foo_caller_contract"
     );
 
     // Calls the contract that calls the `FooContract` contract, also just
@@ -67,7 +66,7 @@ async fn test_reverting_transaction() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/revert_transaction_error"
+        "packages/fuels/tests/contracts/revert_transaction_error"
     );
 
     let response = contract_instance
@@ -85,7 +84,7 @@ async fn test_multiple_read_calls() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/multiple_read_calls"
+        "packages/fuels/tests/contracts/multiple_read_calls"
     );
 
     let contract_methods = contract_instance.methods();
@@ -105,103 +104,11 @@ async fn test_multiple_read_calls() -> Result<(), Error> {
 }
 
 #[tokio::test]
-async fn test_gas_errors() -> Result<(), Error> {
-    let mut wallet = WalletUnlocked::new_random(None);
-    let number_of_coins = 1;
-    let amount_per_coin = 1_000_000;
-    let coins = setup_single_asset_coins(
-        wallet.address(),
-        BASE_ASSET_ID,
-        number_of_coins,
-        amount_per_coin,
-    );
-
-    let (provider, _) = setup_test_provider(coins.clone(), vec![], None).await;
-    wallet.set_provider(provider);
-
-    setup_contract_test!(
-        contract_instance,
-        None,
-        "packages/fuels/tests/behaviors/contract_test"
-    );
-
-    // Test running out of gas. Gas price as `None` will be 0.
-    let gas_limit = 100;
-    let contract_instace_call = contract_instance
-        .methods()
-        .initialize_counter(42) // Build the ABI call
-        .tx_params(TxParameters::new(None, Some(gas_limit), None));
-
-    //  Test that the call will use more gas than the gas limit
-    let gas_used = contract_instace_call
-        .estimate_transaction_cost(None)
-        .await?
-        .gas_used;
-    assert!(gas_used > gas_limit);
-
-    let response = contract_instace_call
-        .call() // Perform the network call
-        .await
-        .expect_err("should error");
-
-    let expected = "Provider error: gas_limit(";
-    assert!(response.to_string().starts_with(expected));
-
-    // Test for insufficient base asset amount to pay for the transaction fee
-    let response = contract_instance
-        .methods()
-        .initialize_counter(42) // Build the ABI call
-        .tx_params(TxParameters::new(Some(100_000_000_000), None, None))
-        .call()
-        .await
-        .expect_err("should error");
-
-    let expected = "Provider error: Response errors; enough coins could not be found";
-    assert!(response.to_string().starts_with(expected));
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_call_param_gas_errors() -> Result<(), Error> {
-    setup_contract_test!(
-        contract_instance,
-        wallet,
-        "packages/fuels/tests/behaviors/contract_test"
-    );
-
-    // Transaction gas_limit is sufficient, call gas_forwarded is too small
-    let contract_methods = contract_instance.methods();
-    let response = contract_methods
-        .initialize_counter(42)
-        .tx_params(TxParameters::new(None, Some(1000), None))
-        .call_params(CallParameters::new(None, None, Some(1)))
-        .call()
-        .await
-        .expect_err("should error");
-
-    let expected = "Revert transaction error: OutOfGas, receipts:";
-    assert!(response.to_string().starts_with(expected));
-
-    // Call params gas_forwarded exceeds transaction limit
-    let response = contract_methods
-        .initialize_counter(42)
-        .tx_params(TxParameters::new(None, Some(1), None))
-        .call_params(CallParameters::new(None, None, Some(1000)))
-        .call()
-        .await
-        .expect_err("should error");
-
-    let expected = "Provider error: gas_limit(";
-    assert!(response.to_string().starts_with(expected));
-    Ok(())
-}
-
-#[tokio::test]
 async fn test_multi_call() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/contract_test"
+        "packages/fuels/tests/contracts/contract_test"
     );
 
     let contract_methods = contract_instance.methods();
@@ -222,40 +129,11 @@ async fn test_multi_call() -> Result<(), Error> {
 }
 
 #[tokio::test]
-async fn test_gas_forwarded_defaults_to_tx_limit() -> Result<(), Error> {
-    setup_contract_test!(
-        contract_instance,
-        wallet,
-        "packages/fuels/tests/behaviors/contract_test"
-    );
-
-    let gas_limit = 225883;
-    let response = contract_instance
-        .methods()
-        .initialize_counter(42)
-        .tx_params(TxParameters::new(None, Some(gas_limit), None))
-        .call()
-        .await?;
-
-    let gas_forwarded = response
-        .receipts
-        .iter()
-        .find(|r| matches!(r, Receipt::Call { .. }))
-        .unwrap()
-        .gas()
-        .unwrap();
-
-    assert_eq!(gas_limit, gas_forwarded);
-
-    Ok(())
-}
-
-#[tokio::test]
 async fn test_contract_call_fee_estimation() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/contract_test"
+        "packages/fuels/tests/contracts/contract_test"
     );
 
     let gas_price = 100_000_000;
@@ -293,7 +171,7 @@ async fn contract_call_has_same_estimated_and_used_gas() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/contract_test"
+        "packages/fuels/tests/contracts/contract_test"
     );
 
     let tolerance = 0.0;
@@ -319,7 +197,7 @@ async fn mutl_call_has_same_estimated_and_used_gas() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/contract_test"
+        "packages/fuels/tests/contracts/contract_test"
     );
 
     let contract_methods = contract_instance.methods();
@@ -349,7 +227,7 @@ async fn contract_method_call_respects_maturity() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/transaction_block_height"
+        "packages/fuels/tests/contracts/transaction_block_height"
     );
 
     let call_w_maturity = |call_maturity| {
@@ -371,7 +249,7 @@ async fn test_auth_msg_sender_from_sdk() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/auth_testing_contract"
+        "packages/fuels/tests/contracts/auth_testing_contract"
     );
 
     // Contract returns true if `msg_sender()` matches `wallet.address()`.
@@ -385,117 +263,12 @@ async fn test_auth_msg_sender_from_sdk() -> Result<(), Error> {
     Ok(())
 }
 
-//-----TODO: sort
-#[tokio::test]
-async fn test_amount_and_asset_forwarding() -> Result<(), Error> {
-    setup_contract_test!(
-        contract_instance,
-        wallet,
-        "packages/fuels/tests/behaviors/token_ops"
-    );
-    let contract_id = contract_instance.get_contract_id();
-    let contract_methods = contract_instance.methods();
-
-    let mut balance_response = contract_methods
-        .get_balance(contract_id.into(), contract_id.into())
-        .call()
-        .await?;
-    assert_eq!(balance_response.value, 0);
-
-    contract_methods.mint_coins(5_000_000).call().await?;
-
-    balance_response = contract_methods
-        .get_balance(contract_id.into(), contract_id.into())
-        .call()
-        .await?;
-    assert_eq!(balance_response.value, 5_000_000);
-
-    let tx_params = TxParameters::new(None, Some(1_000_000), None);
-    // Forward 1_000_000 coin amount of base asset_id
-    // this is a big number for checking that amount can be a u64
-    let call_params = CallParameters::new(Some(1_000_000), None, None);
-
-    let response = contract_methods
-        .get_msg_amount()
-        .tx_params(tx_params)
-        .call_params(call_params)
-        .call()
-        .await?;
-
-    assert_eq!(response.value, 1_000_000);
-
-    let call_response = response
-        .receipts
-        .iter()
-        .find(|&r| matches!(r, Receipt::Call { .. }));
-
-    assert!(call_response.is_some());
-
-    assert_eq!(call_response.unwrap().amount().unwrap(), 1_000_000);
-    assert_eq!(call_response.unwrap().asset_id().unwrap(), &BASE_ASSET_ID);
-
-    let address = wallet.address();
-
-    // withdraw some tokens to wallet
-    contract_methods
-        .transfer_coins_to_output(1_000_000, contract_id.into(), address.into())
-        .append_variable_outputs(1)
-        .call()
-        .await?;
-
-    let asset_id = AssetId::from(*contract_id.hash());
-    let call_params = CallParameters::new(Some(0), Some(asset_id), None);
-    let tx_params = TxParameters::new(None, Some(1_000_000), None);
-
-    let response = contract_methods
-        .get_msg_amount()
-        .tx_params(tx_params)
-        .call_params(call_params)
-        .call()
-        .await?;
-
-    assert_eq!(response.value, 0);
-
-    let call_response = response
-        .receipts
-        .iter()
-        .find(|&r| matches!(r, Receipt::Call { .. }));
-
-    assert!(call_response.is_some());
-
-    assert_eq!(call_response.unwrap().amount().unwrap(), 0);
-    assert_eq!(
-        call_response.unwrap().asset_id().unwrap(),
-        &AssetId::from(*contract_id.hash())
-    );
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_get_gas_used() -> Result<(), Error> {
-    setup_contract_test!(
-        contract_instance,
-        wallet,
-        "packages/fuels/tests/behaviors/contract_test"
-    );
-
-    let gas_used = contract_instance
-        .methods()
-        .initialize_counter(42)
-        .call()
-        .await?
-        .gas_used;
-
-    assert!(gas_used > 0);
-    Ok(())
-}
-
 #[tokio::test]
 async fn test_large_return_data() -> Result<(), Error> {
     setup_contract_test!(
         contract_instance,
         wallet,
-        "packages/fuels/tests/behaviors/large_return_data"
+        "packages/fuels/tests/contracts/large_return_data"
     );
 
     let contract_methods = contract_instance.methods();
@@ -537,5 +310,131 @@ async fn test_large_return_data() -> Result<(), Error> {
             255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255
         ])
     );
+    Ok(())
+}
+
+#[tokio::test]
+async fn can_handle_function_called_new() -> anyhow::Result<()> {
+    setup_contract_test!(
+        contract_instance,
+        wallet,
+        "packages/fuels/tests/contracts/contract_test"
+    );
+
+    let response = contract_instance.methods().new().call().await?.value;
+
+    assert_eq!(response, 12345);
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_contract_setup_macro_deploy_with_salt() -> Result<(), Error> {
+    // ANCHOR: contract_setup_macro_multi
+    // The first wallet name must be `wallet`
+    setup_contract_test!(
+        foo_contract_instance,
+        wallet,
+        "packages/fuels/tests/contracts/foo_contract"
+    );
+    let foo_contract_id = foo_contract_instance.get_contract_id();
+
+    // The macros that want to use the `wallet` have to set
+    // the wallet name to `None`
+    setup_contract_test!(
+        foo_caller_contract_instance,
+        None,
+        "packages/fuels/tests/contracts/foo_caller_contract"
+    );
+    let foo_caller_contract_id = foo_caller_contract_instance.get_contract_id();
+
+    setup_contract_test!(
+        foo_caller_contract_instance2,
+        None,
+        "packages/fuels/tests/contracts/foo_caller_contract"
+    );
+    let foo_caller_contract_id2 = foo_caller_contract_instance2.get_contract_id();
+
+    // Because we deploy with salt, we can deploy the same contract multiple times
+    assert_ne!(foo_caller_contract_id, foo_caller_contract_id2);
+
+    // The first contract can be called because they were deployed on the same provider
+    let bits = *foo_contract_id.hash();
+    let res = foo_caller_contract_instance
+        .methods()
+        .call_foo_contract(Bits256(bits), true)
+        .set_contracts(&[foo_contract_id.clone()]) // Sets the external contract
+        .call()
+        .await?;
+    assert!(res.value);
+
+    let res = foo_caller_contract_instance2
+        .methods()
+        .call_foo_contract(Bits256(bits), true)
+        .set_contracts(&[foo_contract_id.clone()]) // Sets the external contract
+        .call()
+        .await?;
+    assert!(res.value);
+    // ANCHOR_END: contract_setup_macro_multi
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_wallet_getter() -> Result<(), Error> {
+    setup_contract_test!(
+        contract_instance,
+        wallet,
+        "packages/fuels/tests/contracts/contract_test"
+    );
+
+    assert_eq!(contract_instance.get_wallet().address(), wallet.address());
+    //`get_contract_id()` is tested in
+    // async fn test_contract_calling_contract() -> Result<(), Error> {
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_connect_wallet() -> anyhow::Result<()> {
+    // ANCHOR: contract_setup_macro_manual_wallet
+    let config = WalletsConfig::new(Some(2), Some(1), Some(DEFAULT_COIN_AMOUNT));
+
+    let mut wallets = launch_custom_provider_and_get_wallets(config, None).await;
+    let wallet = wallets.pop().unwrap();
+    let wallet_2 = wallets.pop().unwrap();
+
+    setup_contract_test!(
+        contract_instance,
+        None,
+        "packages/fuels/tests/contracts/contract_test"
+    );
+    // ANCHOR_END: contract_setup_macro_manual_wallet
+
+    // pay for call with wallet
+    let tx_params = TxParameters::new(Some(10), Some(10000), None);
+    contract_instance
+        .methods()
+        .initialize_counter(42)
+        .tx_params(tx_params)
+        .call()
+        .await?;
+
+    // confirm that funds have been deducted
+    let wallet_balance = wallet.get_asset_balance(&Default::default()).await?;
+    assert!(DEFAULT_COIN_AMOUNT > wallet_balance);
+
+    // pay for call with wallet_2
+    contract_instance
+        .with_wallet(wallet_2.clone())?
+        .methods()
+        .initialize_counter(42)
+        .tx_params(tx_params)
+        .call()
+        .await?;
+
+    // confirm there are no changes to wallet, wallet_2 has been charged
+    let wallet_balance_second_call = wallet.get_asset_balance(&Default::default()).await?;
+    let wallet_2_balance = wallet_2.get_asset_balance(&Default::default()).await?;
+    assert_eq!(wallet_balance_second_call, wallet_balance);
+    assert!(DEFAULT_COIN_AMOUNT > wallet_2_balance);
     Ok(())
 }
