+++ packages/fuels/tests/contracts/msg_amount/Forc.toml
