+++ packages/fuels/tests/contracts/msg_amount/src/main.sw
