+++ packages/fuels/tests/contracts/multiple_read_calls/Forc.toml
