+++ packages/fuels/tests/contracts/multiple_read_calls/src/main.sw
