+++ packages/fuels/tests/contracts/revert_transaction_error/src/main.sw
