+++ packages/fuels/tests/contracts/token_ops/Forc.toml
