+++ packages/fuels/tests/contracts/token_ops/src/main.sw
